move DataG3Module out of __init__
